2/12
I figured out how to fill a caesar shift cipher. I need to solve the
message still though. I can use sort of the same thing to make the type
in. From there, I made it work depending on what you know. I need to
finish the other general one, and then actually solve it. I don't know
if I want to make a separate class for the actual solve. I started the
monoalph organized grid, but it's a bit harder than I thought. I did
the monoalph grid, but there are a couple of flaws. I need to start
figuring out terminal, and inputting text and exporting text. So then
that'll probably be all I have remaining to do. However, I think it'll
take me a long time to figure that stuff out. Additionally, I'm not
sure if I'm using OOP. I'm not sure where my other .java file is, it's
in the folder, but
I didn't change anything today, so maybe that's why it's not here.